Issue #17: Update dependencies in order fix vulnerabilities
