+++ nbactions.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <actions>
-        <action>
-            <actionName>CUSTOM-dependency:tree</actionName>
-            <displayName>dependency:tree</displayName>
-            <goals>
-                <goal>dependency:tree</goal>
-            </goals>
-        </action>
-    </actions>
+    <action>
+        <actionName>CUSTOM-dependency:tree</actionName>
+        <displayName>dependency:tree</displayName>
+        <goals>
+            <goal>dependency:tree</goal>
+        </goals>
+    </action>
+</actions>
