+++ src/main/bvmui/package.json
@@ -8,17 +8,17 @@
   },
   "dependencies": {
     "axios": "^0.19.0",
-    "lodash": "4.17.11",
+    "lodash": "4.17.15",
     "vue": "2.6.10",
-    "vue-router": "^3.0.6",
+    "vue-router": "^3.0.7",
     "vuex": "3.1.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "3.8.0",
-    "@vue/cli-service": "3.8.4",
-    "sass": "1.21.0",
+    "@vue/cli-plugin-babel": "3.9.2",
+    "@vue/cli-service": "3.9.3",
+    "sass": "1.22.7",
     "sass-loader": "7.1.0",
     "vue-template-compiler": "2.6.10",
-    "webpack": "4.35.0"
+    "webpack": "4.36.1"
   }
 }
